Modified retry limit and sleep time to account for wildfire behavior
Wildfire analysis can take anywhere from 1 second to 15 minutes.  A retry limit of 20 with a sleep of 30 will only allow the cb-wildfire-connector does not account for the full 15 minute period it could take to analyze a binary and queries the api more frequently than necessary.  A retry of 15 and sleep of 60sec is more gentle on the api and accounts for the full time period it can take to analyze a binary.